WIP: NanoProcesses => Actin
TODO: We're now (in theory) building and resolving dependencies.

We need to ensure actors get added to the process pool.
We need to resolve dependencies on all Singletons on startup.
We need to then use the same technique to create instances in scenes and resolve their dependencies.

Then lots of tests.