Fix: See details
Types marked with [Instance] which have a static parent can reuse the static parent if build is being run a second time because they were marked with instance.

The reality is that there is no reason to run build again, as such a type can never actually be created in a scene (Scenes may only use flexible parents), and a type with a static parent must have a parent. However, rather than enforcing this and building a Chesterton's fence, we're just reusing the existing parent type.