Fixes bug when manually adding singleton.
This applies when the director is being used to avoid static references, and the singleton is dynamically added after the normal startup procedure. Because no internal calls are made to the singleton's instantiator, its internal instance is never set. The result is that calls to Instantiator.HasSingletonInstance return false, even though calling GetSingleton internally will return the object.

The simplest way to resolve this is to also check for Instantiator.IsRootSingleton  before deciding that the object is not a singleton.

The commited test fails before this change is made, and succeeds afterwards.